refact(Users): reorganize files
