+++ src/components/user/account-wrapper.jsx
@@ -4,13 +4,11 @@ import { connect } from 'react-redux';
 import { Panel } from 'react-bootstrap';
 
 import { addFlashMessage } from 'services/flash-messages';
-import { updateAccount as updateAccountAction } from 'actions/users.js';
-import { prepareSubmissionError } from 'utils/errors';
-
 import {
-  selectUserInitialValues,
-  selectUserId
-} from 'selectors/users.js';
+  updateAccount as updateAccountAction,
+  selectUserInitialValues, selectUserId,
+} from 'services/users';
+import { prepareSubmissionError } from 'utils/errors';
 
 import AccountForm from './account-form.jsx';
