+++ src/services/auth/saga.js
@@ -2,7 +2,7 @@ import { takeEvery, call, put } from 'redux-saga/effects';
 
 import { getCookies, setCookies } from 'utils/cookies';
 import { postToken } from 'api/token.js';
-import { getCurrentUser } from 'api/users.js';
+import { getCurrentUser } from 'services/users/api.js';
 import { getOrganization } from 'services/organizations/api.js';
 
 import {
